NlpChatbot/requirements.txt
streamlit
faiss-cpu
transformers
sentence-transformers
PyMuPDF
PyPDF2
beautifulsoup4
requests
torch
